+++ Installation/CHANGES.md
@@ -76,7 +76,7 @@ Release date: June 2020
   was imposed to the user. Other concepts have been clarified to reflect the fact that the API uses this particular class.
 
 ### Convex_hull_3
-- Added a new overload for `VertexListGraph`.
+- A new overload for `convex_hull_3()` that takes a model of `VertexListGraph` has been added.
 
 [Release 5.0](https://github.com/CGAL/cgal/releases/tag/releases%2FCGAL-5.0)
 -----------